Add client certificates only if any provided
Not all .NET Framework versions that claim to be netstandard2.0 
compliant implement methods necessary for the 
ClientCertificates property to be accessed and result in a 
MissingMethodException being thrown. 
This conditional will only attempt to access that property if 
additional client certificates are provided. 
Additionally, the only usage by HttpSender to create a 
THttpClientTransport does not ever pass in additional certificates 
so this will always skip over the property access. 
As a result, this library can be used on older .NET Frameworks that 
otherwise implement the netstandard2.0 framework.

Signed-off-by: rwkarg <rwkarg@gmail.com>