Adding identifying info to warning logging
It is difficult to identify which is the offending span from the current log line. This change is to include information to easily identify which span needs to be addressed; either through a common operation triggers multiple warning log lines or through the specific trace/span id that can be investigated through the dashboard.